Remove old and unused blocking ConsulKV client
This implements WhiskConfig using ConsulClient instead of the old ConsulKV, making it unreferenced and therefore removed.
Await.result is tolerated in WhiskConfig as it is only executed once per component.
